+++ src/name.cc
@@ -4,7 +4,7 @@
 //
 // ##Copyright##
 // 
-// Copyright (C) 2000-2004
+// Copyright (C) 2000-2009 
 // School of Information Technology and Electrical Engineering
 // The University of Queensland
 // Australia 4072
@@ -14,9 +14,6 @@
 // The Qu-Prolog System and Documentation  
 // 
 // COPYRIGHT NOTICE, LICENCE AND DISCLAIMER.
-// 
-// Copyright 2000-2004 by The University of Queensland, 
-// Queensland 4072 Australia
 // 
 // Permission to use, copy and distribute this software and associated
 // documentation for any non-commercial purpose and without fee is hereby 
@@ -393,6 +390,6 @@ Thread::psi_codes_number(Object *& object1, Object *& object2)
     if (has_dot || has_e)
       object2 = TheHeap().newDouble(atof(numstr));
     else
-      object2 = TheHeap().newInteger(atoi(numstr));
+      object2 = TheHeap().newInteger(atol(numstr));
     return(RV_SUCCESS);
 }
