+++ src/write.cc
@@ -3,7 +3,7 @@
 //
 // ##Copyright##
 // 
-// Copyright (C) 2000-2004
+// Copyright (C) 2000-2009 
 // School of Information Technology and Electrical Engineering
 // The University of Queensland
 // Australia 4072
@@ -13,9 +13,6 @@
 // The Qu-Prolog System and Documentation  
 // 
 // COPYRIGHT NOTICE, LICENCE AND DISCLAIMER.
-// 
-// Copyright 2000-2004 by The University of Queensland, 
-// Queensland 4072 Australia
 // 
 // Permission to use, copy and distribute this software and associated
 // documentation for any non-commercial purpose and without fee is hereby 
@@ -329,7 +326,7 @@ Thread::psi_write_var(Object *& object1, Object *& object2)
       // No name.
       //
      *(stream) << "_";
-     *(stream) << (int)(reinterpret_cast<heapobject*>(var) - heap.getBase());
+     *(stream) << (long)(reinterpret_cast<heapobject*>(var) - heap.getBase());
     }
   return(RV_SUCCESS);
 }
